refactor: split chainrules-core into traits crate + AD engine crate
chainrules-core now contains only core AD traits (Differentiable,
ReverseRule, ForwardRule) like Julia's ChainRulesCore.jl.

New chainrules crate contains the AD engine (TrackedTensor, DualTensor,
pullback, hvp, Gradients, PullbackPlan, HvpResult) and re-exports
chainrules-core. Follows Rust foo-core/foo convention.

tenferro-einsum now depends on chainrules instead of chainrules-core.
tenferro-tensor keeps chainrules-core dep (only uses Differentiable trait).
